добавлен новый модуль mas.js
