+++ js/data.js
@@ -1,8 +1,7 @@
-import {getRandomArrayElement, getRandomInteger} from './rand.js';
-
 const description = ['пляж', 'лес', 'море', 'юг', 'оеан', 'лето', 'весна', 'котики', 'будни',
   'работа', 'лайфхаки', 'блог', 'прайс-лист', 'ночь', 'эстетика',
-  'книги', 'еда', 'зоопарк', 'театр', 'прогулка', 'путешествие', 'парк', 'закат', 'музыка', 'хобби' ];
+  'книги', 'еда', 'зоопарк', 'театр', 'прогулка', 'путешествие', 'парк', 'закат', 'музыка', 'хобби',
+];
 const messageCommentator = [
   'Всё отлично!',
   'В целом всё неплохо. Но не всё.',
@@ -18,20 +17,4 @@ const nameCommentator = [
 
 const count = 25;
 
-const createComment = (indexx) => ({
-  id: indexx + 1,
-  avatar: `img/avatar-${ getRandomInteger(1, 6) } .svg`,
-  message: `${getRandomArrayElement(messageCommentator)}`,
-  name: `${getRandomArrayElement(nameCommentator)}`,
-});
-
-const createUsers = (index) => ({
-  id: index + 1,
-  url: `photos/${ index + 1 }.jpg` ,
-  description: `${getRandomArrayElement(description)}`,
-  likes: getRandomInteger(1, 200),
-  comments: Array.from({ length: getRandomInteger(1, 30) }, (__, indexx) => createComment(indexx)),
-});
-const finalMas = Array.from({length:count}, (__, index) => createUsers(index));
-
-export {finalMas};
+export {description, messageCommentator, nameCommentator, count};
