+++ js/main.js
@@ -1,3 +1,3 @@
-import { finalMas } from './data';
+import { finalMas } from '.js/mas.js';
 // eslint-disable-next-line no-console
 console.log(finalMas);
